--DSC - PIFSC Oracle Developer Environment: Initial Working Version--
Forked the PIFSC Oracle Developer Environment (ODE) repository to customize the development environment for the DSC database schema

Updated deployment_scripts/prepare_docker_project.sh to retrieve the DSC project files from the git repository and place them into the corresponding locations to be included in the docker image when it's built.

Updated docker/src/run_db_app_deployment.sh to comment out the code that waits for APEX to be installed since there is no APEX app for the DSC project.  Added code to create and deploy the DSC schema.

Updated README.md to specify the information for the customized DSC ODE